Nightly updates
Bug fix for scanning loop: Include corrupt sfo files to the cache, the index lua file needs to find them or else a onelua scan is triggered. LPP will exclude the games anyway as the category is "UNK".

Scanning method also updated to use quickScanGames to mirror the scanning in LPP to further reduce chance of scanning loop.